ui/requirements.txt
streamlit==1.32.0
requests==2.31.0
matplotlib==3.8.2
websockets==12.0
opencv-python==4.9.0.80
numpy==1.26.4
Pillow==10.2.0
aiohttp==3.9.3
